+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Add `merchantInitiatedRequestType` and `priorAuthenticationId` to `ThreeDSecureLookupRequest`
 * Add `foreignRetailer` to `Transaction` and `TransactionRequest`
+* Add `internationalPhone` to `Address` and `Customer`
 
 ## 3.34.0
 * Add domains parameter support to `ClientTokenRequest`